requirements.txt
Flask
psycopg2-binary

#blinker==1.9.0
#click==8.1.3  # Use a version compatible with Python 3.9
#colorama==0.4.6
#Flask==3.1.1
#psycopg2==2.9.10
#itsdangerous==2.2.0
#Jinja2==3.1.6
#MarkupSafe==3.0.2
#Werkzeug==3.1.3
